+++ lib/src/metrics_factory.dart
@@ -1,5 +1,6 @@
 import 'package:meta/meta.dart';
 
+import 'metrics/cyclomatic_complexity/cyclomatic_complexity_metric.dart';
 import 'metrics/lines_of_code_metric.dart';
 import 'metrics/maximum_nesting_level/maximum_nesting_level_metric.dart';
 import 'metrics/metric.dart';
@@ -9,6 +10,8 @@ import 'metrics/weight_of_class_metric.dart';
 import 'models/entity_type.dart';
 
 final _implementedMetrics = <String, Metric Function(Map<String, Object>)>{
+  CyclomaticComplexityMetric.metricId: (config) =>
+      CyclomaticComplexityMetric(config: config),
   LinesOfCodeMetric.metricId: (config) => LinesOfCodeMetric(config: config),
   MaximumNestingLevelMetric.metricId: (config) =>
       MaximumNestingLevelMetric(config: config),